replaced XoopsConfirm

+++ admin/resources.php
@@ -426,19 +426,19 @@ switch ($op) {
             }
         } else {
             if ('delete_all' == $op) {
-                $xoopsconfirm = new Common\XoopsConfirm(
+                $customConfirm = new Common\Confirm(
                     ['ok' => 1, 'res_pro_id' => $proId, 'op' => 'delete_all'],
                     $_SERVER['REQUEST_URI'],
                     \sprintf(\_AM_WGTRANSIFEX_RESOURCES_SURE_DELETEALL, $projectsObj->getVar('pro_name'))
                 );
             } else {
-                $xoopsconfirm = new Common\XoopsConfirm(
+                $customConfirm = new Common\Confirm(
                     ['ok' => 1, 'res_id' => $resId, 'res_pro_id' => $proId, 'op' => 'delete'],
                     $_SERVER['REQUEST_URI'],
                     \sprintf(\_AM_WGTRANSIFEX_FORM_SURE_DELETE, $resourcesObj->getVar('res_slug'))
                 );
             }
-            $form = $xoopsconfirm->getFormXoopsConfirm();
+            $form = $customConfirm->getFormConfirm();
             $GLOBALS['xoopsTpl']->assign('form', $form->render());
         }
         break;
